requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6
langchain>=0.0.344
langchain-community>=0.0.13
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
openai>=1.3.3
python-dotenv>=1.0.0
pypdf>=3.17.1
sqlalchemy>=2.0.23